backend/requirements.txt
# === Core Application Libraries ===
python-dotenv
cryptography

# === API Framework ===
fastapi
uvicorn[standard]
pydantic
python-multipart

# === Manually Resolved Dependencies ===
moviepy==1.0.3
tensorflow
opencv-contrib-python

# === AI & Analysis Libraries ===
fer
transformers
torch
librosa
sentencepiece
onnx
onnxruntime
soundfile
pydub

# === External API SDKs ===
groq
deepgram-sdk==2.12.0
edge-tts